# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Pi Zero 2W Medicine Tracker API
  description: |
    RESTful API for medicine tracking and management on Pi Zero 2W.
    This API provides endpoints for managing medicines, tracking adherence,
    and configuring application settings.
  version: 1.0.0
  contact:
    name: Pi Zero 2W Project
  license:
    name: MIT
servers:
  - url: http://localhost:5000/api/v1
    description: Development server
  - url: http://{hostname}:5000/api/v1
    description: Pi Zero 2W device server
    variables:
      hostname:
        default: pizero.local
        description: Hostname or IP address of the Pi Zero 2W device
...
paths:
...
  /config:
    get:
      tags:
        - Configuration
      summary: Get All Configuration
      description: Get entire application configuration
      operationId: getAllConfig
      responses:
        '200':
          description: All configuration sections
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    additionalProperties: true
                  meta:
                    $ref: '#/components/schemas/Meta'
        '500':
          description: Error reading configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Configuration
      summary: Update All Configuration
      description: Replace entire application configuration
      operationId: updateAllConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Configuration updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  message:
                    type: string
                  meta:
                    $ref: '#/components/schemas/Meta'
        '400':
          description: Invalid configuration data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error saving configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /config/{section}:
    get:
      tags:
        - Configuration
      summary: Get Configuration Section
      description: Get specific configuration section
      operationId: getConfigSection
      parameters:
        - name: section
          in: path
          required: true
          description: Configuration section name (weather, mbta, disney, flights, pomodoro, forbidden, medicine, menu, system, display)
          schema:
            type: string
            enum:
              - weather
              - mbta
              - disney
              - flights
              - pomodoro
              - forbidden
              - medicine
              - menu
              - system
              - display
      responses:
        '200':
          description: Configuration section data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  meta:
                    $ref: '#/components/schemas/Meta'
        '404':
          description: Section not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error reading configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Configuration
      summary: Replace Configuration Section
      description: Replace entire configuration section
      operationId: replaceConfigSection
      parameters:
        - name: section
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Configuration section updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  message:
                    type: string
                  meta:
                    $ref: '#/components/schemas/Meta'
        '400':
          description: Invalid data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Section not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error saving configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      tags:
        - Configuration
      summary: Update Configuration Section (Partial)
      description: Partial update of configuration section
      operationId: patchConfigSection
      parameters:
        - name: section
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Configuration section updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  message:
                    type: string
                  meta:
                    $ref: '#/components/schemas/Meta'
        '400':
          description: Invalid data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Section not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error saving configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    Meta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: When the response was generated
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              description: Additional error details
        meta:
          $ref: '#/components/schemas/Meta'
...